Agregamos carpetas por paginas y componentes
