--- my-app/src/components/login.jsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const Login = () => {
-    return (
-        <div>
-            <h1>Login</h1>
-        </div>
-    )
-}
-
-export default Login
