+++ my-app/src/pages/login/login.jsx
@@ -1,0 +1,14 @@
+import React from 'react'
+import './login.css'
+const Login = () => {
+    return (
+        <div className="container loginContainer" >
+            <div className="loginLogo"></div>
+            <h1 className="loginText">INICIA SESIÓN</h1>
+            <input type="email" id="email" class="input" placeholder="Ingresa tu email"/>
+            <input type="password" id="emailCreateAccount" class="input" placeholder="Ingresa tu contraseña"/>
+            <input type="button" id="btnLogin" class="loginButton" value="INICIA SESIÓN"/>
+        </div>
+    )
+}
+export default Login
