+++ my-app/src/pages/new-order/new-order.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import Header from './header'
+import Header from '../../components/header/header'
 
 const NewOrder = () => {
     return (
