+++ my-app/src/style/App.css
@@ -5,26 +5,10 @@
   min-height: 100vh;
   color: #fff;
 }
-header{
-  width: 100%;
-  height: 70px;
-  border-bottom:#fff solid 2px;
-  padding: 30px;
-
-}
-.headerLogo{
-  background: url('../assets/img/logo-BQ.svg')no-repeat;  
-  width: auto;
-  height: 75px;
-}
 
 .green{
   background-color: #0CA789;
 }
-.button{
-  width: 30%;
-  height: 30vh; 
-  border-radius: 50px;
-  color: #fff;
-  text-decoration: none;
-}
+
+/* Estilos An */
+
